REST API for books authors and publishers is finished

+++ src/books/books.controller.ts
@@ -7,11 +7,14 @@ import {
   Param,
   Delete,
   UseGuards,
+  HttpException,
+  HttpStatus,
 } from '@nestjs/common';
 import { BooksService } from './books.service';
 import { CreateBookDto } from './dto/create-book.dto';
+import { CreateBookBulk } from './dto/create-book-bulk.dto';
 import { UpdateBookDto } from './dto/update-book.dto';
-import { JWTAuthGuard } from '../auth/jwt-auth.guard';
+import { JWTAuthGuard } from '../guards/jwt-auth.guard';
 
 @Controller('books')
 export class BooksController {
@@ -23,6 +26,12 @@ export class BooksController {
     return this.booksService.create(createBookDto);
   }
 
+  @Post('/bulk')
+  @UseGuards(JWTAuthGuard)
+  createBulk(@Body() createBookDto: CreateBookBulk) {
+    return this.booksService.createBulk(createBookDto);
+  }
+
   @Get()
   @UseGuards(JWTAuthGuard)
   findAll() {
@@ -31,19 +40,34 @@ export class BooksController {
 
   @Get(':id')
   @UseGuards(JWTAuthGuard)
-  findOne(@Param('id') id: string) {
+  async findOne(@Param('id') id: string) {
+    if (isNaN(+id))
+      throw new HttpException(
+        `id - '${id}' is Not a Number`,
+        HttpStatus.BAD_REQUEST,
+      );
     return this.booksService.findOne(+id);
   }
 
   @Patch(':id')
   @UseGuards(JWTAuthGuard)
   update(@Param('id') id: string, @Body() updateBookDto: UpdateBookDto) {
+    if (isNaN(+id))
+      throw new HttpException(
+        `id - '${id}' is Not a Number`,
+        HttpStatus.BAD_REQUEST,
+      );
     return this.booksService.update(+id, updateBookDto);
   }
 
   @Delete(':id')
   @UseGuards(JWTAuthGuard)
   remove(@Param('id') id: string) {
+    if (isNaN(+id))
+      throw new HttpException(
+        `id - '${id}' is Not a Number`,
+        HttpStatus.BAD_REQUEST,
+      );
     return this.booksService.remove(+id);
   }
 }
